+++ packages/pieces/community/wedof/src/lib/actions/registration-folders/cancel-registration-folder.ts
@@ -1,15 +1,13 @@
 import { HttpMethod, httpClient } from '@activepieces/pieces-common';
-import { wedofAuth } from '../..';
+import { wedofAuth } from '../../..';
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { wedofCommon } from '../common/wedof';
-import dayjs from 'dayjs';
+import { wedofCommon } from '../../common/wedof';
 
-export const declareRegistrationFolderTerminated = createAction({
+export const cancelRegistrationFolder = createAction({
   auth: wedofAuth,
-  name: 'declareRegistrationFolderTerminated',
-  displayName: "Passer un dossier de formation à l'état : sortie de formation",
-  description:
-    "Change l'état d'un dossier de formation vers : sortie de formation",
+  name: 'cancelRegistrationFolder',
+  displayName: 'Annuler le dossier de formation',
+  description: 'Annuler le dossier de formation',
   props: {
     externalId: Property.ShortText({
       displayName: 'N° du dossier de formation',
@@ -17,14 +15,9 @@ export const declareRegistrationFolderTerminated = createAction({
         'Sélectionner la propriété {externalId} du dossier de formation',
       required: true,
     }),
-    date: Property.DateTime({
-      displayName: 'Sortie de formation le',
-      description: 'Date au format YYYY-MM-DD.',
-      required: false,
-    }),
     code: Property.Dropdown({
-      displayName: 'Raison de la sortie de formation',
-      description: 'Sélectionner la raison de sortie de formation',
+      displayName: "Raison de l'annulation du dossier de formation",
+      description: "Sélectionner la raison de l'annulation",
       required: true,
       refreshers: ['auth'],
       refreshOnSearch: false,
@@ -39,8 +32,7 @@ export const declareRegistrationFolderTerminated = createAction({
           await httpClient.sendRequest({
             method: HttpMethod.GET,
             url:
-              wedofCommon.baseUrl +
-              '/registrationFoldersReasons?type=terminated',
+              wedofCommon.baseUrl + '/registrationFoldersReasons?type=canceled',
             headers: {
               'Content-Type': 'application/json',
               'X-Api-Key': auth as string,
@@ -58,20 +50,16 @@ export const declareRegistrationFolderTerminated = createAction({
         };
       },
     }),
-    absenceDuration: Property.Number({
-      displayName: "durée d'absence",
-      description:
-        "La durée d'une éventuelle absence en heures. 0 si aucune absence.",
+    description: Property.LongText({
+      displayName: 'Description',
+      description: " Texte expliquant les raisons de l'annulation",
       required: false,
     }),
   },
   async run(context) {
     const message = {
-      date: context.propsValue.date
-        ? dayjs(context.propsValue.date).format('YYYY-MM-DD')
-        : null,
       code: context.propsValue.code,
-      absenceDuration: context.propsValue.absenceDuration,
+      description: context.propsValue.description,
     };
 
     return (
@@ -81,7 +69,7 @@ export const declareRegistrationFolderTerminated = createAction({
           wedofCommon.baseUrl +
           '/registrationFolders/' +
           context.propsValue.externalId +
-          '/terminate',
+          '/cancel',
         body: message,
         headers: {
           'Content-Type': 'application/json',
