+++ packages/pieces/community/wedof/src/lib/actions/registration-folders/declare-registration-folder-intraining.ts
@@ -1,7 +1,7 @@
 import { HttpMethod, httpClient } from '@activepieces/pieces-common';
-import { wedofAuth } from '../..';
+import { wedofAuth } from '../../..';
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { wedofCommon } from '../common/wedof';
+import { wedofCommon } from '../../common/wedof';
 import dayjs from 'dayjs';
 
 export const declareRegistrationFolderIntraining = createAction({
