+++ packages/pieces/community/wedof/src/lib/actions/registration-folders/refuse-registration-folder.ts
@@ -1,7 +1,7 @@
 import { HttpMethod, httpClient } from '@activepieces/pieces-common';
-import { wedofAuth } from '../..';
+import { wedofAuth } from '../../..';
 import { createAction, Property } from '@activepieces/pieces-framework';
-import { wedofCommon } from '../common/wedof';
+import { wedofCommon } from '../../common/wedof';
 
 export const refuseRegistrationFolder = createAction({
   auth: wedofAuth,
