+++ packages/pieces/community/wedof/src/lib/common/wedof.ts
@@ -1,8 +1,8 @@
 import { Property } from '@activepieces/pieces-framework';
 
 export const wedofCommon = {
-  baseUrl: 'https://www.wedof.fr/api',
-  host: 'https://www.wedof.fr/api',
+  baseUrl: 'https://staging.wedof.fr/api',
+  host: 'https://staging.wedof.fr/api',
 
   state: Property.StaticMultiSelectDropdown({
     displayName: 'Etat du dossier de formation',
